requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
gunicorn==21.2.0
Flask-Session==0.5.0
cachelib==0.10.2